% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_parse.R
\name{cook_adjacent}
\alias{cook_adjacent}
\title{FUNCTION_TITLE}
\usage{
cook_adjacent(cook_open_data)
}
\arguments{
\item{cook_open_data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map2}}
 \code{\link[tidyr]{unnest}}
}
