% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_ballotpedia.R
\name{cook_ballotpedia}
\alias{cook_ballotpedia}
\title{FUNCTION_TITLE}
\usage{
cook_ballotpedia(cook_clean)
}
\arguments{
\item{cook_clean}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
