% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_parse.R
\name{cook_open}
\alias{cook_open}
\title{FUNCTION_TITLE}
\usage{
cook_open(cook_data)
}
\arguments{
\item{cook_data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
