% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_plot.R
\name{cook_plot}
\alias{cook_plot}
\title{FUNCTION_TITLE}
\usage{
cook_plot(dat)
}
\arguments{
\item{dat}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
