% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adj_mat.R
\name{create_adj_mat}
\alias{create_adj_mat}
\title{FUNCTION_TITLE}
\usage{
create_adj_mat()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
District shp file to adj mat
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[geojsonio]{geojson_read}},\code{\link[geojsonio]{geojson_sp}}
 \code{\link[rgeos]{gTouches}}
}
