% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_districts.R
\name{fetch_districts}
\alias{fetch_districts}
\title{FUNCTION_TITLE}
\usage{
fetch_districts(raw_data, idx)
}
\arguments{
\item{raw_data}{PARAM_DESCRIPTION}

\item{idx}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
