% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_history.R
\name{fetch_history}
\alias{fetch_history}
\title{FUNCTION_TITLE}
\usage{
fetch_history(State, district)
}
\arguments{
\item{State}{PARAM_DESCRIPTION}

\item{district}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
 \code{\link[rvest]{html_table}}
}
