% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_spreadsheet.R
\name{fetch_spreadsheet}
\alias{fetch_spreadsheet}
\title{Cook spreadsheet}
\usage{
fetch_spreadsheet(cook_html)
}
\arguments{
\item{cook_html}{response, object returned from GET call of the cook google spreadsheet.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Tinyurl address of cook google spreadsheet http://tinyurl.com/y9jxsgmt
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
