% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_history.R
\name{parse_history}
\alias{parse_history}
\title{FUNCTION_TITLE}
\usage{
parse_history(dat)
}
\arguments{
\item{dat}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
