% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_spreadsheet.R
\name{fetch_spreadsheet}
\alias{fetch_spreadsheet}
\title{Cook spreadsheet}
\usage{
fetch_spreadsheet()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
